Add IsCross property to distinguish cross-build
SkipBuildingInstaller is set as true,
because in cross-build environment we can't test installer file.
In cross-build, it doesn't have additional shared framework so that it is set to false.
And it also should be installed the shared framework archive that matches the RID,
so UsePortableLinuxSharedFramework is also set to false.
DISABLE_CROSSGEN is not used currently, so it is removed.